Backend/requirements.txt
pip
autopep8
Flask
gunicorn
Werkzeug
fastapi
uvicorn
pydantic
python-dotenv
datasets
qdrant-client
langchain
langchain-google-genai
langchain-community
langgraph
langchain-tavily
nemoguardrails
httpx
aiofiles
asyncio-mqtt
websockets
json-schema
dataclasses-json